Separated TaskConfig into its own file.
