+++ source/core/src/main/com/csse3200/game/components/player/inventory/TargetDummy.java
@@ -3,7 +3,6 @@ package com.csse3200.game.components.player.inventory;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.math.GridPoint2;
 import com.csse3200.game.ai.tasks.AITaskComponent;
-import com.csse3200.game.areas.BaseRoom;
 import com.csse3200.game.areas.EnemyRoom;
 import com.csse3200.game.components.npc.NPCConfigComponent;
 import com.csse3200.game.components.npc.attack.MeleeAttackComponent;
@@ -12,6 +11,7 @@ import com.csse3200.game.components.tasks.ChargeTask;
 import com.csse3200.game.components.tasks.ChaseTask;
 import com.csse3200.game.entities.Entity;
 import com.csse3200.game.entities.configs.NPCConfigs;
+import com.csse3200.game.entities.configs.TaskConfig;
 import com.csse3200.game.entities.factories.DeployableItemFactory;
 import com.csse3200.game.services.ServiceLocator;
 
@@ -65,7 +65,7 @@ public class TargetDummy extends UsableItem {
             if (entity.getComponent(AITaskComponent.class) != null) {
 
                 NPCConfigs.NPCConfig config = entity.getComponent(NPCConfigComponent.class).config;
-                NPCConfigs.NPCConfig.TaskConfig tasks = config.tasks;
+                TaskConfig tasks = config.tasks;
 
                 MeleeAttackComponent meleeAttack = entity.getComponent(MeleeAttackComponent.class);
                 RangeAttackComponent rangedAttack = entity.getComponent(RangeAttackComponent.class);
