+++ source/core/src/test/com/csse3200/game/components/tasks/RunAwayTaskTest.java
@@ -5,6 +5,7 @@ import com.csse3200.game.ai.tasks.AITaskComponent;
 import com.csse3200.game.components.CombatStatsComponent;
 import com.csse3200.game.entities.Entity;
 import com.csse3200.game.entities.configs.NPCConfigs;
+import com.csse3200.game.entities.configs.TaskConfig;
 import com.csse3200.game.extensions.GameExtension;
 import com.csse3200.game.physics.PhysicsService;
 import com.csse3200.game.physics.components.PhysicsComponent;
@@ -44,7 +45,7 @@ class RunAwayTaskTest {
         target.setPosition(2f, 2f);
 
         // Use a configuration object for RunAwayTask
-        NPCConfigs.NPCConfig.TaskConfig.RunAwayTaskConfig config = new NPCConfigs.NPCConfig.TaskConfig.RunAwayTaskConfig();
+        TaskConfig.RunAwayTaskConfig config = new TaskConfig.RunAwayTaskConfig();
         config.runSpeed = 3;
         config.stopDistance = 5;
         config.maxRunTime = 10;
@@ -80,7 +81,7 @@ class RunAwayTaskTest {
         target.setPosition(2f, 2f);
 
         // Use a configuration object for RunAwayTask
-        NPCConfigs.NPCConfig.TaskConfig.RunAwayTaskConfig config = new NPCConfigs.NPCConfig.TaskConfig.RunAwayTaskConfig();
+        TaskConfig.RunAwayTaskConfig config = new TaskConfig.RunAwayTaskConfig();
         config.runSpeed = 3;
         config.stopDistance = 5;
         config.maxRunTime = 10;
@@ -120,7 +121,7 @@ class RunAwayTaskTest {
         target.setPosition(20f, 20f); // Target out of view distance
 
         // Use a configuration object for RunAwayTask
-        NPCConfigs.NPCConfig.TaskConfig.RunAwayTaskConfig config = new NPCConfigs.NPCConfig.TaskConfig.RunAwayTaskConfig();
+        TaskConfig.RunAwayTaskConfig config = new TaskConfig.RunAwayTaskConfig();
         config.runSpeed = 3;
         config.stopDistance = 5;
         config.maxRunTime = 10;
@@ -151,7 +152,7 @@ class RunAwayTaskTest {
         target.setPosition(2f, 2f);
 
         // Use a configuration object for RunAwayTask
-        NPCConfigs.NPCConfig.TaskConfig.RunAwayTaskConfig config = new NPCConfigs.NPCConfig.TaskConfig.RunAwayTaskConfig();
+        TaskConfig.RunAwayTaskConfig config = new TaskConfig.RunAwayTaskConfig();
         config.runSpeed = 20;
         config.stopDistance = 5;
         config.maxRunTime = 10;
@@ -183,7 +184,7 @@ class RunAwayTaskTest {
         target.setPosition(2f, 2f);
 
         // Use a configuration object for RunAwayTask
-        NPCConfigs.NPCConfig.TaskConfig.RunAwayTaskConfig config = new NPCConfigs.NPCConfig.TaskConfig.RunAwayTaskConfig();
+        TaskConfig.RunAwayTaskConfig config = new TaskConfig.RunAwayTaskConfig();
         config.runSpeed = 3;
         config.stopDistance = 5;
         config.maxRunTime = 10;
